requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings>=2.1.0
google-generativeai==0.3.2
edge-tts==6.1.0
pillow>=10.0.0
moviepy>=1.0.0
manim>=0.19.0
celery==5.3.4
redis==5.0.1
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
structlog==23.2.0
httpx==0.25.2
numpy>=1.24.0
opencv-python>=4.8.0
scipy>=1.11.0